// src/modules/interfaces/filterParam.ts
export type FilterFormParam = {
  param: string;
  nameParam: string;
  nameParamForAttribute: string;
  nameParamTranslate: string;
  svg: string;
};
export type FilterColorParam = {
  param: string;
  nameParam: string;
  nameParamForAttribute: string;
  svg: string;
  nameParamTranslate: string;
};
export type FilterSizeParam = {
  param: string;
  nameParam: string;
  nameParamForAttribute: string;
  nameParamTranslate: string;
  svg: string;
};